ml-service/model/requirements.txt
# NyayaSetu ML — Python Dependencies

# Core ML Framework
torch>=2.0.0
transformers>=4.35.0
scikit-learn>=1.3.0

# Data Processing
numpy>=1.24.0
pandas>=2.0.0

# Audio Processing (for ASR)
librosa>=0.10.0
soundfile>=0.12.0
# openai-whisper>=20231117  # Optional: for better ASR

# API & Utilities
requests>=2.31.0
pydantic>=2.0.0

# Optional: For GPU Training
# torch --index-url https://download.pytorch.org/whl/cu118

# Optional: For Model Quantization
# bitsandbytes>=0.41.0
huggingface_hub
python-dotenv
